Delete all "fuzzy" tags by brute force.
This appears to work! Note that I am __not__ replacing emtpy msgstrs with `XXX` as previously discussed, and this commit is for French only.

So the deal is: `compilemessages` assumes anything marked `fuzzy` is wrong and __skips it__. We should be able to recompile .mo files without this hacky string replacement with the `--use-fuzzy` flag