Fix input values overlapping with % on indicator target form
`.periodic-target .input-value.` in _custom.scss was overriding `.input-symbol-percent .input-value` in _mc_tola_forms.scss, which is why the right padding was incorrect.

Switching the order such that the forms CSS comes after custom CSS matches the ordering that used to exist in app.css, which is why this works on production.