A workaround to keep filter text after indicator selection
This is a workaround as it fundamentally breaks a part of this React component, namely it will *not* update available options in the select if new options are passed in through the props. If we ever run across a case where the select options need to change over time, and also a filter is wanted, this decision will need to be revisited.

The background to this is Bootstrap Multiselect does not provide an API to easily set the filter string.

Fixes an issue where the first element isn't selectable after a selection is made