Remove usage of FontAwesomeIcon where icons are not dynamic
There were some JS errors being thrown as a result of using this, unless an icon was added to the FA library. We are using 2 FA versions here, one is included in base.html, and one is included in NPM. Try to prefer using the base.html version in the hopes the React/NPM version can be ripped out in the future.

Also fixed a few instances of class="" being used in JSX