Fix character encoding issue on IPTT excel export
`l10n_date_long` returns a utf-8 encoded str, which when shoved into u"".format(), causes an encoding error when the date contains unicode chars. Fix this by converting the dates back to a unicode type. This was done as opposed to doing "".format() to keep the inputs consistent with all of the other unicode strings being fed in.

Closes #1657